feat: add react-router and initial routes
