+++ services/client/src/components/__tests__/App.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { shallow } from 'enzyme'
+
+import App from '../../App'
+
+test('App renders without crashing', () => {
+  const wrapper = shallow(<App />)
+})
